+++ CHANGELOG.md
@@ -33,6 +33,10 @@ and this project adheres to [Semantic Versioning][Semantic Versioning].
 
 ## [0.0.1] - 2025-04-17
 
+### Added
+
+- link between doc properties and the SaveFileView form auto fills when user presses search btn
+
 ### Changed
 
 - divided the `ConfigHelper.vb` into `PathHelper` and `ListHelper`